finalize FR_O parser with historical fetching
