+++ config/zones/GF.yaml
@@ -12,6 +12,5 @@ capacity:
 contributors:
   - PETILLON-Sebastien
 parsers:
-  price: FR_O.fetch_price
   production: FR_O.fetch_production
 timezone: Europe/Paris
