+++ config/zones/GP.yaml
@@ -14,6 +14,5 @@ capacity:
 contributors:
   - PETILLON-Sebastien
 parsers:
-  price: FR_O.fetch_price
   production: FR_O.fetch_production
 timezone: Europe/Paris
